requirements.txt
backcall==0.1.0
decorator==4.3.0
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
numpy==1.22.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
Pillow==9.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyclust==0.2.0
pyclustering==0.8.1
Pygments==2.15.0
scikit-learn==0.23.2
scipy==1.10.0
simplegeneric==0.8.1
six==1.11.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.23.3
traitlets==4.3.2
wcwidth==0.1.7
treelib
pandas
pydicom
scikit-image